# Fix ColoredConsole performance issues