--- a/src/util/console-color.ts
+++ b/src/util/console-color.ts
@@ -6,6 +6,7 @@ interface ConsoleState {
   foregroundColor: string | null;
   backgroundColor: string | null;
   carriageReturn: boolean;
+  lines: string[];
   secret: boolean;
 }
 
@@ -18,6 +19,7 @@ export class ColoredConsole {
     foregroundColor: null,
     backgroundColor: null,
     carriageReturn: false,
+    lines: [],
     secret: false,
   };
 
@@ -27,25 +29,12 @@ export class ColoredConsole {
     return this.targetElement.innerText;
   }
 
-  addLine(line: string) {
+  processLine(line: string): Element {
     const re = /(?:\x1b|\\033)(?:\[(.*?)[@-~]|\].*?(?:\x07|\x07\\))/g;
     let i = 0;
 
-    if (this.state.carriageReturn) {
-      if (line !== "\n") {
-        // don't remove if \r\n
-        this.targetElement.removeChild(this.targetElement.lastChild!);
-      }
-      this.state.carriageReturn = false;
-    }
-
-    if (line.includes("\r")) {
-      this.state.carriageReturn = true;
-    }
-
     const lineSpan = document.createElement("span");
     lineSpan.classList.add("line");
-    this.targetElement.appendChild(lineSpan);
 
     const addSpan = (content: string) => {
       if (content === "") return;
@@ -178,16 +167,51 @@ export class ColoredConsole {
         }
       }
     }
+    addSpan(line.substring(i));
+    return lineSpan;
+  }
+
+  processLines() {
     const atBottom =
       this.targetElement.scrollTop >
       this.targetElement.scrollHeight - this.targetElement.offsetHeight - 50;
+    const prevCarriageReturn = this.state.carriageReturn;
+    const fragment = document.createDocumentFragment();
 
-    addSpan(line.substring(i));
+    if (this.state.lines.length == 0) {
+      return;
+    }
+
+    for (const line of this.state.lines) {
+      if (this.state.carriageReturn && line !== "\n") {
+        if (fragment.childElementCount) {
+          fragment.removeChild(fragment.lastChild!);
+        }
+      }
+      fragment.appendChild(this.processLine(line));
+      this.state.carriageReturn = line.includes("\r");
+    }
+
+    if (prevCarriageReturn && this.state.lines[0] !== "\n") {
+      this.targetElement.replaceChild(fragment, this.targetElement.lastChild!);
+    } else {
+      this.targetElement.appendChild(fragment);
+    }
+
+    this.state.lines = [];
 
     // Keep scroll at bottom
     if (atBottom) {
       this.targetElement.scrollTop = this.targetElement.scrollHeight;
     }
+  }
+
+  addLine(line: string) {
+    // processing of lines is deferred for performance reasons
+    if (this.state.lines.length == 0) {
+      setTimeout(() => this.processLines(), 0);
+    }
+    this.state.lines.push(line);
   }
 }
 
@@ -203,6 +227,7 @@ export const coloredConsoleStyles = `
     line-height: 1.45;
     border-radius: 3px;
     white-space: pre-wrap;
+    will-change: scroll-position;
     overflow-wrap: break-word;
     color: #ddd;
   }